# perf: optimize retrieving the set of equally best paths 